completely rework printing approach to add csv writing capability
There is now an "Outputter" class that handles print() commands,
the writing to a text file, and collecting all the printed text into
a list that can be saved as a CSV file. The new class is set up to be
used as a context manager, i.e. in 'with' statements, to ensure that
the file object for text saving is closed